added option to link dotfiles

+++ bootstrap.sh
@@ -2,20 +2,49 @@
 
 cd $(dirname "$0") || exit 1
 
+USAGE="Usage: $(basename $0) [-hl]"
+
+while getopts :hl opt; do
+    case $opt in
+        h)
+            echo $USAGE
+            exit 0
+            ;;
+        l)
+            link_dotfiles=1
+            ;;
+        \?)
+            echo >&2 $USAGE
+            exit 1
+            ;;
+    esac
+done
+
 dotfiles=( bash_profile bashrc inputrc nethackrc screenrc vimrc )
 umask 0077
 
 for df in "${dotfiles[@]}"; do
     dst_df="${HOME}/.${df}"
-    if [[ -L ${dst_df} ]]; then
-        echo >&2 "Removing symlink .${df}"
-        rm "${dst_df}"
-    fi
 
-    cp "${df}" "${dst_df}"
-    # || { echo "Could not copy .${df} into home folder"; exit 1 }
+    if [[ $link_dotfiles ]]; then
+        # if requested to create links to dotfiles, does so
+        ln -sf "$(pwd)/${df}" "${dst_df}"
+    else
+        # otherwise removes links (if any)...
+        if [[ -L ${dst_df} ]]; then
+            echo >&2 "Removing symlink .${df}"
+            rm "${dst_df}"
+        fi
+
+        # .. and copies dotfiles
+        cp "${df}" "${dst_df}"
+        # || { echo "Could not copy .${df} into home folder"; exit 1 }
+    fi
 done
 
+shift $(expr $OPTIND - 1)
+
+# vim stuff
 for vd in "~/.vim/backups" "~/.vim/swaps"; do
     if [[ ! -d $vd ]]; then
         mkdir -p "$vd"
@@ -23,6 +52,7 @@ for vd in "~/.vim/backups" "~/.vim/swaps"; do
 done
 
 if [[ $(uname) = 'Darwin' ]]; then
+    # if running os x
     if which brew >/dev/null; then
         # taps
         brew tap homebrew/games >/dev/null
